Support panel

Create the standard DfE R-Shiny support and feedback dashboard panel.

Usage

support_panel(
  team_email = "",
  repo_name = "",
  ees_publication = TRUE,
  publication_name = NULL,
  publication_slug = "",
  alt_href = NULL,
  form_url = NULL,
  cookie_status_output = "cookie_status"
)

Arguments

team_email

Your team e-mail address, must be a education.gov.uk email

repo_name

The repository URL, must be a valid URL for the dfe-analytical-services GitHub area

ees_publication

Whether the parent publication is hosted on Explore Education Statistics

publication_name

The parent publication name

publication_slug

The parent publication slug on Explore Education Statistics

alt_href

Alternative link to the parent publication (if not hosted on Explore Education Statistics)

form_url

URL for a feedback form for the dashboard

cookie_status_output

Name of cookie status output object, often "cookie_status"

Value

a standardised panel for a public R Shiny dashboard in DfE

Examples

support_panel(
  team_email = "my.team@education.gov.uk",
  repo_name = "https://github.com/dfe-analytical-services/my-repo",
  publication_name = "My publication title",
  publication_slug = "my-publication-title",
  form_url = "www.myform.com",
  cookie_status_output = "cookie_status"
)
#> <div class="tab-pane" title="Support and feedback" data-value="support_panel">
#>   <div id="main" class="govuk-width-container">
#>     <main class="govuk-main-wrapper">
#>       <div class="govuk-grid-row">
#>         <div class="col-sm-12">
#>           <h1>Support and feedback</h1>
#>           <h2>Give us feedback</h2>
#>           <p>
#>             This dashboard is a new service that we are developing. If you
#>               have any feedback or suggestions for improvements, please submit
#>               them using our 
#>             <a href="www.myform.com">feedback form</a></p>
#>           <p>
#>             Alternatively, if you spot any errors or bugs while using this dashboard, please
#>               screenshot and email them to 
#>             <a href="mailto:my.team@education.gov.uk">my.team@education.gov.uk</a>.
#>           </p>
#>           <h2>Find more information on the data</h2>
#>           <p>
#>             The parent statistical release of this dashboard, along with
#>               methodological information,
#>               is available at the following link: 
#>             <a href="https://explore-education-statistics.service.gov.uk/find-statistics/my-publication-title">My publication title</a>. The statistical release provides additional 
#>             <a href="https://explore-education-statistics.service.gov.uk/find-statistics/my-publication-title/data-guidance">data guidance</a> and 
#>             <a href="https://explore-education-statistics.service.gov.uk/find-statistics/my-publication-title#explore-data-and-files">tools to access and interogate the underling data</a> contained in this dashboard.
#>           </p>
#>           <h2>Contact us</h2>
#>           <p>
#>             If you have questions about the dashboard or data within it,
#>             please contact us at 
#>             <a href="mailto:my.team@education.gov.uk">my.team@education.gov.uk</a></p>
#>           <h2>See the source code</h2>
#>           <p>
#>             The source code for this dashboard is available in our 
#>             <a href="https://github.com/dfe-analytical-services/my-repo">GitHub repository</a>.
#>           </p>
#>           <h2>Use of cookies</h2>
#>           <p>To better understand the reach of our dashboard tools,
#>             this site uses cookies to identify numbers of unique users
#>             as part of Google Analytics.</p>
#>           <div id="cookie_status" class="shiny-text-output"></div>
#>           <button id="cookie_consent_clear" type="button" class="btn btn-default action-button">Reset cookie consent</button>
#>         </div>
#>       </div>
#>     </main>
#>   </div>
#> </div>