Custom disconnect message

Create the html overlay panel to appear when RSConnect disconnects

Usage

custom_disconnect_message(
  refresh = "Refresh page",
  links = NULL,
  publication_name = NULL,
  publication_link = NULL
)

Arguments

refresh

the text to appear that will refresh the page when clicked

links

A list of mirrors or alternative links to the dashboard

publication_name

The parent publication name

publication_link

The link to the publication on Explore Education Statistics

Value

A html overlay panel that appears when RSConnect disconnects for a public R Shiny dashboard in DfE

Examples

custom_disconnect_message(
  refresh = "Refresh page",
  links = c(
    "https://department-for-education.shinyapps.io/dfe-shiny-template/",
    "https://department-for-education.shinyapps.io/dfe-shiny-template-overflow/" # nolint: [line_length_linter]
  ),
  publication_name = "Explore Education Statistics Publication",
  publication_link =
    "https://explore-education-statistics.service.gov.uk/find-statistics/pupil-attendance-in-schools" # nolint: [line_length_linter]
)
#> <script>$(function() {  $(document).on('shiny:disconnected', function(event) {    $('#custom-disconnect-dialog').show();    $('#ss-overlay').show();  })});</script>
#> <div id="custom-disconnect-dialog" style="display: none !important;">
#>   <div id="ss-connect-refresh">
#>     <p>You've lost connection to the dashboard server - please try
#>                refreshing the page:</p>
#>     <p>
#>       <a id="ss-reload-link" href="#" onclick="window.location.reload(true);">Refresh page</a>
#>     </p>
#>     <p>
#>       If this persists, you can also view the dashboard at one of our
#>             mirror sites:
#>       <p>
#>         <a href="https://department-for-education.shinyapps.io/dfe-shiny-template/">Site 1</a>
#>          - 
#>         <a href="https://department-for-education.shinyapps.io/dfe-shiny-template-overflow/">Site 2</a>
#>       </p>
#>     </p>
#>     <p>
#>       All the data used in this dashboard can also be viewed or
#>             downloaded via the 
#>       <a href="https://explore-education-statistics.service.gov.uk/find-statistics/pupil-attendance-in-schools">Explore Education Statistics Publication</a>
#>       on Explore Education Statistics.
#>     </p>
#>     <p>
#>       Please contact
#>       <a href="mailto:statistics.development@education.gov.uk">statistics.development@education.gov.uk</a>
#>       with details of any problems with this resource.
#>     </p>
#>   </div>
#> </div>
#> <div id="ss-overlay" style="display: none;"></div>